requirements.txt
neat-python==0.92
numpy==1.19.0
Pillow==7.2.0
pygame==1.9.6
six==1.15.0
